refactor: move ringing-screen Activity customisation from MainActivity.kt into the alarm-manager plugin
MainActivity.kt (under gen/android, generated by `tauri android init`) carried
two hand-written customisations that a fresh regeneration would silently
drop: the onNewIntent -> setIntent() fix and the showWhenLocked/turnScreenOn
keyguard-occlusion backstop added last session. Tauri's own auto-generated
TauriActivity base class already dispatches onNewIntent to every registered
plugin via PluginManager, and each plugin already holds an Activity reference
-- exactly the mechanism DeepLinkPlugin.kt already uses for its own
onNewIntent handling. Move both behaviours into AlarmManagerPlugin's own
load()/onNewIntent(), which is also the natural owner (it already owns
AlarmReceiver/AlarmRingingService, the rest of the native ringing pipeline).
MainActivity.kt now matches Tauri's stock scaffold output exactly.

Also fixes a real gap the move surfaced: showWhenLocked/turnScreenOn were
never reset once set, so they'd stay on for the rest of that Activity
instance's life -- a later, unrelated app open while locked (without killing
the app first) could bypass the lock screen too. stopRinging() is the one
command every path out of the ringing screen (dismiss, snooze, silence
timeout) already calls, so it's the natural place to clear them back.
